Organizing groups inside of bullet lists
I took the liberty to organize HH groups into bullet lists. It looks far more estetic this way (in my own opinion). I haven't done anything else special in any way, so if you think that the previous design was better, feel free to deny this as a pull commit.