Update README.md
Group name has been changed to "HH 3D". We're trying to make it more broad in order to foster growth in the group. This is now a subgroup of Hackathon Hackers devoted to all things with length, width, and height. 3D printing, design, modeling, animation, art, software, hardware, and related topics.